+++ gulp/gulp.build.js
@@ -1,0 +1,15 @@
+/**
+ * Created by hzou on 3/3/16.
+ */
+
+
+module.exports = {
+  init: init
+};
+
+function init(gulp, plugins, config, _) {
+  gulp.task('build', function (done) {
+    var tasks = ['html', 'fonts', 'images', 'scripts', 'templates', 'assets', 'sass', 'favicon', 'version'];
+    plugins.seq('clean', tasks, done);
+  });
+}