+++ gulp/gulp.components.js
@@ -1,0 +1,20 @@
+/**
+ * Created by hzou on 3/3/16.
+ */
+
+const fs    = require('fs');
+const path  = require('path');
+const gutil = require('gulp-util');
+
+module.exports = {
+  init: init
+};
+
+function init(gulp, plugins, config, _) {
+
+  gulp.task('component.json', function (done) {
+    gulp.src(['src/components/foundation/default/component.json']).
+      pipe(gulp.dest('gulp'));
+  });
+}
+