+++ src/components/foundation/base/modules/isc.components.module.js
@@ -1,0 +1,26 @@
+/**
+ * Created by douglasgoodman on 11/21/14.
+ */
+
+(function () {
+  'use strict';
+
+  angular.module('isc.components',
+    [
+      'isc.common',
+      // third party modules
+      'angularUtils.directives.dirPagination',
+      '720kb.datepicker',
+      'foundation',
+      'angucomplete-alt',
+      'formly',
+      'isc.forms',
+      'isc.directives',
+      'isc.table',
+      'isc.ddpTable',
+      'isc.notification',
+      'isc.spinner',
+      'iscNavContainer'
+    ]);
+
+})();