+++ src/components/foundation/base/modules/isc.spinner/iscSpinner.module.js
@@ -1,0 +1,12 @@
+/**
+ * Created by hzou on 1/2/16.
+ */
+
+(function () {
+  'use strict';
+
+  angular
+    .module('isc.spinner', []);
+
+})();
+